Adds sign-in functionality
Implements a new sign-in page with input fields for email and password.

It uses the `useAuth` hook to handle authentication and redirects the user to their intended destination after successful login, or the homepage if no destination is specified.

The sign-in page includes a visual design with logos and illustrations.