Enhances dashboard with user experience improvements
Improves the dashboard with several UI enhancements.
It adds a sparkes SVG image and a text input element
to suggest to the user to ask Tymmar to fill in the
week's work hours.

Adds also a message to remind the user there are unsaved
changes, promoting data persistence.

Disables week period edition and displays a tertiary-colored
label when closed.

Increments input number step from 0.25 to 1 to accept only
integer numbers.